+++ babylon60/extensions/evolution/ouroboros_omega.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 # [C5-REAL] Exergy-Maximized
 """
 Ouroboros-Omega: The Central Autopoietic Metabolism of CORTEX.
@@ -38,7 +39,7 @@ class DiagnosisMatrix:
     used_imports: set[str] = field(default_factory=set)
     dead_interfaces: set[str] = field(default_factory=set)
     blocking_calls: list[str] = field(default_factory=list)
-    entropy_score: float = 0.0
+    entropy_score: Decimal = 0.0
 
 
 class OuroborosOmega:
@@ -187,7 +188,7 @@ class OuroborosOmega:
                         p0_report.critical_count,
                         p0_report.high_count,
                     )
-                except Exception as e:
+                except Exception as e:  # noqa: BLE001
                     logger.warning("Phase 1.5 [P0 Scan] Failed: %s", e)
 
             tree = ast.parse(self.original_source)
@@ -222,7 +223,7 @@ class OuroborosOmega:
             try:
                 ast.parse(mutated_source)  # Syntax
                 compile(mutated_source, filename="<ast>", mode="exec")  # Bytecode
-            except Exception as e:
+            except Exception as e:  # noqa: BLE001
                 logger.error("Phase 5 [Verification] Failed syntax/bytecode: %s", e)
                 return {"status": "ROLLED_BACK", "reason": str(e)}
 
@@ -271,7 +272,7 @@ class OuroborosOmega:
                     )
                 except ImportError as e:
                     logger.warning("Could not dispatch remote mutation: %s", e)
-                except Exception as e:
+                except Exception as e:  # noqa: BLE001
                     logger.error("Terminal State 4 dispatch failed: %s", e)
 
             # COMMIT
@@ -292,7 +293,7 @@ class OuroborosOmega:
                 result["p0_report"] = p0_report.to_dict()
             return result
 
-        except Exception as e:
+        except Exception as e:  # noqa: BLE001
             logger.exception("Apoptosis: Unhandled exception during cycle.")
             return {"status": "ROLLED_BACK", "reason": str(e)}
