+++ babylon60/extensions/perception/metabolism.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 # [C5-REAL] Exergy-Maximized
 """
 Perception Layer 2: Metabolic Scaling.
@@ -35,7 +36,7 @@ class MetabolicObserver:
         self._consecutive_low_exergy = 0
 
     def calculate_intensity(
-        self, event_count: int, avg_diff_size: float, exergy_score: float | None = None
+        self, event_count: int, avg_diff_size: float, exergy_score: Decimal | None = None
     ) -> str:
         """
         Determine the target intensity based on activity and exergy.
